requirements.txt
fastapi
pydantic
uvicorn[standard]
python-dotenv
httpx
pyjwt
python-multipart
pymongo
motor
apscheduler
pytest 